refactor(earnings-report): getNextEarningsReport를 lib에서 api.ts로 이동 (순수 함수 레이어 위반 해소)
- getNextEarningsReport(DB + FMP fetch side-effect)를 entities/earnings-report/lib/에서 api.ts로 이동
- lib/은 이제 순수 함수 isEarningsReportStale만 보유
- barrel(index.ts)은 동일하게 export → 기존 호출자 영향 없음
- test: 기존 단위 테스트를 prototype-spy mock 전략으로 재설계
