+++ src/entities/earnings-report/api.ts
@@ -9,6 +9,11 @@ import { NEON_TRANSIENT_RETRY } from '@/shared/db/isNeonTransientError';
 import { earningsReports } from '@/shared/db/schema';
 import type { SiglensDatabase } from '@/shared/db/types';
 import { withRetry } from '@/shared/lib/withRetry';
+import { isEarningsReportStale } from './lib/isEarningsReportStale';
+import { getFundamentalDataProvider } from '@/shared/api/fmp/getFundamentalDataProvider';
+import { todayKstIsoDate } from '@/shared/lib/dateKey';
+
+export const EARNINGS_REPORT_FMP_LIMIT = 5;
 
 export interface EarningsReportUpsertInput {
     symbol: string;
@@ -304,4 +309,37 @@ function isPresent(value: number | null): boolean {
 
 function hasEstimateValue(row: EarningsReportDataRow): boolean {
     return row.epsEstimated !== null || row.revenueEstimated !== null;
+}
+
+/**
+ * Returns the next upcoming earnings entry for `symbol`, refreshing from FMP if
+ * the DB has no data or the last fetch is older than 24 hours.
+ * Used by analysis actions that run independently of the news page visit.
+ *
+ * DB·FMP·`Date.now()` side effect를 포함하는 use-case이므로 순수 함수 전용 레이어
+ * (lib/)가 아니라 api.ts(server-only repository/API 계층)에 둔다. staleness 판정은
+ * 순수 함수 `isEarningsReportStale`(lib/)에 위임하고, `Date.now()`는 여기
+ * (infrastructure 경계)에서 주입한다.
+ */
+export async function getNextEarningsReport(
+    symbol: string,
+    db: SiglensDatabase
+): Promise<EarningsCalendarItem | null> {
+    const repo = new DrizzleEarningsReportsRepository(db);
+    const fetchedAt = await repo.getLatestFetchedAt(symbol);
+
+    if (isEarningsReportStale(fetchedAt, Date.now())) {
+        try {
+            const client = getFundamentalDataProvider();
+            const reports = await client.getEarningsReports(
+                symbol,
+                EARNINGS_REPORT_FMP_LIMIT
+            );
+            await repo.upsertMany(reports);
+        } catch {
+            // Best-effort: analysis proceeds without earnings context if FMP fails
+        }
+    }
+
+    return repo.getNextForSymbol(symbol, todayKstIsoDate());
 }
