+++ src/entities/earnings-report/index.ts
@@ -1,12 +1,13 @@
 export {
     DrizzleEarningsReportsRepository,
+    EARNINGS_REPORT_FMP_LIMIT,
     dedupeEarningsReportInputs,
+    getNextEarningsReport,
     toComparisonItems,
     type EarningsReportUpsertInput,
 } from './api';
 
 // lib
-export { getNextEarningsReport } from './lib/nextEarningsReport';
 export {
     EARNINGS_REPORT_STALE_MS,
     isEarningsReportStale,
