+++ src/shared/api/fmp/FakeFundamentalDataProvider.ts
@@ -123,7 +123,7 @@ export class FakeFundamentalDataProvider implements FundamentalDataProvider {
 
     /**
      * Siglens-specific extra (not on the core port) — some injection points
-     * (newsData.ts, nextEarningsReport.ts) call it on the concrete client.
+     * (newsData.ts, earnings-report의 api.ts getNextEarningsReport) call it on the concrete client.
      * Returning an empty list keeps the earnings-report DB upsert a no-op.
      */
     async getEarningsReports(
